initial draft emp apis
